+++ lib/src/radio/sbb_radio_button_list_item.dart
@@ -12,17 +12,16 @@ const double _minListTileHeight = 44.0;
 ///
 /// See also:
 ///
-/// * [SBBRadioButton], which is a part of this widget.
-/// * [SBBSegmentedButton], a widget with semantics similar to [SBBRadioButton].
+/// * [SBBRadio], which is a part of this widget.
+/// * [SBBSegmentedButton], a widget with semantics similar to [SBBRadio].
 /// * [SBBSlider], for selecting a value in a range.
 /// * [SBBCheckboxListItem], [SBBCheckbox] and [SBBSwitch], for toggling a
 /// particular value on or off.
 /// * <https://digital.sbb.ch/de/design-system-mobile-new/elemente/radiobutton>
-@Deprecated('Use SBBRadioListItem instead')
-class SBBRadioButtonListItem<T> extends StatelessWidget {
-  /// Creates a combination of a tile and a [SBBRadioButton].
+class SBBRadioListItem<T> extends StatelessWidget {
+  /// Creates a combination of a tile and a [SBBRadio].
   ///
-  /// The [SBBRadioButtonListItem] itself does not maintain any state. Instead, when the
+  /// The [SBBRadioListItem] itself does not maintain any state. Instead, when the
   /// state of the checkbox changes, the widget calls the [onChanged] callback.
   /// Most widgets that use a checkbox will listen for the [onChanged] callback
   /// and rebuild the checkbox tile with a new [value] to update the visual
@@ -32,7 +31,7 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
   ///
   /// * [value] and [groupValue] together determine whether the radio button is
   ///   selected.
-  /// * [label], the primary text written on [SBBCheckboxListItem].
+  /// * [label], the primary text written on [SBBRadioListItem].
   /// * [onChanged] is called when the user selects this radio button.
   ///
   /// Set the [isLastElement] true for the last item in a list to not show any Divider.
@@ -45,9 +44,8 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
   ///
   /// If [isLoading] is true, a bottom loading indicator will be displayed.
   ///
-  /// Check the [SBBRadioButtonListItem.custom] constructor for a complete customization.
-  @Deprecated('Use SBBRadioListItem instead.')
-  SBBRadioButtonListItem({
+  /// Check the [SBBRadioListItem.custom] constructor for a complete customization.
+  SBBRadioListItem({
     Key? key,
     required T value,
     required T? groupValue,
@@ -76,7 +74,7 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
           radioSemanticLabel: radioSemanticLabel,
         );
 
-  /// Use this in combination with a [SBBGroup] to create a boxed variant of the [SBBRadioButtonListItem].
+  /// Use this in combination with a [SBBGroup] to create a boxed variant of the [SBBRadioListItem].
   ///
   /// ```dart
   /// SBBGroup(
@@ -92,8 +90,8 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
   /// )
   ///
   /// ```
-  @Deprecated('Use SBBRadioListItem instead.')
-  SBBRadioButtonListItem.boxed({
+  ///
+  SBBRadioListItem.boxed({
     Key? key,
     required T value,
     required T? groupValue,
@@ -121,8 +119,7 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
           radioSemanticLabel: radioSemanticLabel,
         );
 
-  @Deprecated('Use SBBRadioListItem instead.')
-  const SBBRadioButtonListItem.custom({
+  const SBBRadioListItem.custom({
     super.key,
     required this.value,
     required this.groupValue,
@@ -146,9 +143,9 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
   /// [groupValue].
   final T? groupValue;
 
-  /// Called when the user selects this [SBBRadioButtonListItem].
+  /// Called when the user selects this [SBBRadioListItem].
   ///
-  /// The [SBBRadioButtonListItem] passes [value] as a parameter to this callback. The list item
+  /// The [SBBRadioListItem] passes [value] as a parameter to this callback. The list item
   /// does not actually change state until the parent widget rebuilds the
   /// list item with the new [groupValue].
   ///
@@ -202,22 +199,82 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
   /// This label does not show in the UI.
   final String? radioSemanticLabel;
 
+  /// Whether [value] of this control can be changed by user interaction.
+  ///
+  /// The control is considered interactive if the [onChanged] callback is
+  /// non-null. If the callback is null, then the control is disabled, and
+  /// non-interactive.
+  bool get _isInteractive => onChanged != null;
+
   @override
   Widget build(BuildContext context) {
-    return SBBRadioListItem.custom(
-      key: key,
-      value: value,
-      groupValue: groupValue,
-      onChanged: onChanged,
-      label: label,
-      allowMultilineLabel: allowMultilineLabel,
-      secondaryLabel: secondaryLabel,
-      isLastElement: isLastElement,
-      leadingIcon: leadingIcon,
-      trailingWidget: trailingWidget,
-      isLoading: isLoading,
-      radioSemanticLabel: radioSemanticLabel,
+    final style = SBBControlStyles.of(context).radioButton;
+    return MergeSemantics(
+      child: Material(
+        color: _resolveBackgroundColor(style),
+        child: Stack(
+          alignment: Alignment.bottomCenter,
+          children: [
+            ConstrainedBox(
+              constraints: const BoxConstraints(minHeight: _minListTileHeight),
+              child: InkWell(
+                onTap: _isInteractive ? () => onChanged!(value) : null,
+                splashColor: style?.listItem?.backgroundColorHighlighted,
+                focusColor: style?.listItem?.backgroundColorHighlighted,
+                highlightColor: SBBColors.transparent,
+                hoverColor: SBBColors.transparent,
+                child: Semantics(
+                  enabled: _isInteractive,
+                  child: _radioTileBody(style),
+                ),
+              ),
+            ),
+            if (!isLastElement) const Divider(),
+            if (isLoading) BottomLoadingIndicator(),
+          ],
+        ),
+      ),
     );
+  }
+
+  Widget _radioTileBody(SBBControlStyle? style) {
+    final Color? resolvedIconColor = _resolveIconColor(style);
+
+    return IconTheme.merge(
+      data: IconThemeData(color: resolvedIconColor),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const SizedBox(width: sbbDefaultSpacing),
+          _NonHittableRadio<T>(
+            value: value,
+            groupValue: groupValue,
+            onChanged: onChanged,
+            radioSemanticLabel: radioSemanticLabel,
+          ),
+          const SizedBox(width: sbbDefaultSpacing * 0.5),
+          if (leadingIcon != null) _LeadingIcon(leadingIcon: leadingIcon),
+          Expanded(
+            child: _TextBody(
+              label: label,
+              isInteractive: _isInteractive,
+              style: style,
+              allowMultilineLabel: allowMultilineLabel,
+              secondaryLabel: secondaryLabel,
+            ),
+          ),
+          if (trailingWidget != null) trailingWidget!,
+        ],
+      ),
+    );
+  }
+
+  Color? _resolveBackgroundColor(SBBControlStyle? style) {
+    return _isInteractive ? style?.listItem?.backgroundColor : style?.listItem?.backgroundColorDisabled;
+  }
+
+  Color? _resolveIconColor(SBBControlStyle? style) {
+    return _isInteractive ? style?.listItem?.iconColor : style?.listItem?.iconColorDisabled;
   }
 
   static Widget? _optionallyButtonedTrailingIcon<T>(
@@ -241,5 +298,102 @@ class SBBRadioButtonListItem<T> extends StatelessWidget {
         ),
       );
     }
+  }
+}
+
+class _TextBody extends StatelessWidget {
+  const _TextBody({
+    required this.label,
+    required bool isInteractive,
+    required this.style,
+    required this.allowMultilineLabel,
+    required this.secondaryLabel,
+  }) : _isInteractive = isInteractive;
+
+  final String label;
+  final bool _isInteractive;
+  final SBBControlStyle? style;
+  final bool allowMultilineLabel;
+  final String? secondaryLabel;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsetsDirectional.only(
+        top: sbbDefaultSpacing * 0.75,
+        bottom: sbbDefaultSpacing * 0.75,
+        end: sbbDefaultSpacing,
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            label,
+            style: _isInteractive ? style?.listItem?.textStyle : style?.listItem?.textStyleDisabled,
+            maxLines: allowMultilineLabel ? null : 1,
+            overflow: allowMultilineLabel ? null : TextOverflow.ellipsis,
+          ),
+          if (secondaryLabel != null)
+            Padding(
+              padding: const EdgeInsetsDirectional.only(
+                top: sbbDefaultSpacing * 0.25,
+              ),
+              child: Text(
+                secondaryLabel!,
+                style:
+                    _isInteractive ? style?.listItem?.secondaryTextStyle : style?.listItem?.secondaryTextStyleDisabled,
+              ),
+            ),
+        ],
+      ),
+    );
+  }
+}
+
+class _LeadingIcon extends StatelessWidget {
+  const _LeadingIcon({required this.leadingIcon});
+
+  final IconData? leadingIcon;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsetsDirectional.only(top: _iconTopPadding, end: sbbDefaultSpacing * .5),
+      child: Icon(leadingIcon),
+    );
+  }
+}
+
+class _NonHittableRadio<T> extends StatelessWidget {
+  const _NonHittableRadio({
+    required this.value,
+    required this.groupValue,
+    required this.onChanged,
+    this.radioSemanticLabel,
+  });
+
+  final T value;
+  final T? groupValue;
+  final ValueChanged<T?>? onChanged;
+  final String? radioSemanticLabel;
+
+  @override
+  Widget build(BuildContext context) {
+    return IgnorePointer(
+      child: ExcludeFocus(
+        child: Padding(
+          padding: const EdgeInsetsDirectional.only(
+            top: sbbDefaultSpacing * 0.75,
+          ),
+          child: SBBRadio(
+            value: value,
+            groupValue: groupValue,
+            onChanged: onChanged,
+            padding: EdgeInsets.zero,
+            semanticLabel: radioSemanticLabel,
+          ),
+        ),
+      ),
+    );
   }
 }